% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_helper_packages.R
\docType{package}
\name{GlmSimulatoR-package}
\alias{GlmSimulatoR}
\alias{GlmSimulatoR-package}
\title{GlmSimulatoR: Creates Ideal Data for Generalized Linear Models}
\description{
Creates ideal data for all distributions in the generalized
    linear model framework.
}
\keyword{internal}
